WIP port to 1.21.8
Compiles, datagen works & game runs, but there are many bugs and crashes

+++ src/main/java/archives/tater/netherarchives/mixin/BlockMixin.java
@@ -1,6 +1,12 @@
 package archives.tater.netherarchives.mixin;
 
 import archives.tater.netherarchives.item.SkisItem;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FluidBlock;
@@ -8,10 +14,6 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Block.class)
 public class BlockMixin {
@@ -20,7 +22,7 @@ public class BlockMixin {
             method = "onLandedUpon",
             at = @At("HEAD"),
             cancellable = true)
-    private void preventFallOnSkis(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance, CallbackInfo ci) {
+    private void preventFallOnSkis(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance, CallbackInfo ci) {
         //noinspection ConstantValue
         if (!((Object) this instanceof FluidBlock)) return;
         if (!(entity instanceof LivingEntity livingEntity) || !SkisItem.canSki(livingEntity, state.getFluidState())) return;
